Split SEDModel.kt
